shell contact algorithm
